requirements.txt
# Core dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.0.0
transformers>=4.30.0

# Generation
openai>=1.0.0

# UI
gradio>=4.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Training
peft>=0.5.0
datasets>=2.14.0

# Utilities
tqdm>=4.65.0
